Fixes #2975: Memberships are deleted when PATCHed
When PATCHing a team, the documentation states that passing users
(memberships) replaces all existing members. But when a user that is
already part of a team is passed, he is removed from that team. This
change fixes that issue by first emptying the team (removing all
memberships) if team members are passed through the patch method. Since
doctrine / the entity manager seems to process the inserts before the
deletes, we need to also manually flush the entity manager before saving
the team again, so we don't get duplicate key sql errors.