Fixes #3495: Quick entry during grace period
During the grace period, new time entries could not be added on the
first day of the lockdown period. This was due to an edge case where the
start time of the timesheet exactly matched the start time of the
lockdown period, and we checked for "greater than" instead of "greater
than or equal to".