Update Util.php
Proposal to fix rate rounding issues on invoices. Since the invoices show the duration as a decimal number with two decimal points.

Example: 
assuming that the hourly rate is 34.34 and the duration is 25 minutes
Current algorithm calculates 1500/3600*34.34 = 14.31 €
proposal algorithm calculates 0.42 (number on the invoice) * 34.34 = 14.42 €

It's dirty but works for me. I don't expect that to be part of the repo